+++ typescript/src/client/evm/evm.types.ts
@@ -39,6 +39,7 @@ export type EvmClientInterface = Omit<
   | "getEvmAccount" // mapped to getAccount
   | "getEvmAccountByName" // mapped to getAccount
   | "getEvmSmartAccount" // mapped to getSmartAccount
+  | "getEvmSmartAccountByName" // mapped to getSmartAccount
   | "getEvmSwapPrice" // mapped to getSwapPrice
   | "createEvmSwapQuote" // mapped to createSwapQuote
   | "getUserOperation"
@@ -56,6 +57,8 @@ export type EvmClientInterface = Omit<
   | "sendEvmTransaction" // mapped to sendTransaction
   | "signEvmTypedData" // mapped to signTypedData
   | "updateEvmAccount" // mapped to updateAccount
+  | "exportEvmAccount"
+  | "exportEvmAccountByName"
 > & {
   createAccount: (options: CreateServerAccountOptions) => Promise<ServerAccount>;
   createSmartAccount: (options: CreateSmartAccountOptions) => Promise<SmartAccount>;
@@ -336,9 +339,11 @@ export interface GetServerAccountOptions {
  */
 export interface GetSmartAccountOptions {
   /** The address of the account. */
-  address: Address;
+  address?: Address;
   /** The owner of the account. */
   owner: Account;
+  /** The name of the account. */
+  name?: string;
 }
 
 /**
@@ -347,6 +352,16 @@ export interface GetSmartAccountOptions {
 export interface GetOrCreateServerAccountOptions {
   /** The name of the account. */
   name: string;
+}
+
+/**
+ * Options for getting an EVM account, or creating one if it doesn't exist.
+ */
+export interface GetOrCreateSmartAccountOptions {
+  /** The name of the account. */
+  name: string;
+  /** The owner of the account. */
+  owner: Account;
 }
 
 /**
@@ -426,6 +441,8 @@ export interface CreateSmartAccountOptions {
   owner: Account;
   /** The idempotency key. */
   idempotencyKey?: string;
+  /** The name of the account. */
+  name?: string;
 }
 
 /**